+++ choicescript/src/util.js
@@ -296,7 +296,13 @@ function logout(callback) {
 function recordLogin(registered, loginId, email, callback) {
   if (initStore()) {
     if (registered) recordEmail(email);
-    window.store.set("login", loginId || 0, function() {safeCall(null, callback);});
+    window.store.set("login", loginId || 0, function() {
+      if (_global.isAndroidApp) {
+        window.receiptRequestCallback = submitReceipts;
+        androidBilling.requestReceipts();
+      }
+      safeCall(null, callback);
+    });
     window.registered = registered;
   } else {
     safeTimeout(callback, 0);
@@ -686,6 +692,7 @@ function getAppId() {
 
 function submitReceipts(receipts, callback) {
   console.log("submitReceipts: " + JSON.stringify(receipts));
+  if (!initStore()) return;
   if (!callback) callback = function(error) {
     if (window.transferPurchaseCallback) {
       window.transferPurchaseCallback(error || "done");
@@ -861,6 +868,7 @@ function restoreGame(state, forcedScene, userRestored, forcedStats, forcedTemps)
     if (!isStateValid(state)) {
         var startupScene = forcedScene ? forcedScene : _global.nav.getStartupScene();
         scene = new Scene(startupScene, _global.stats, _global.nav, {debugMode:_global.debug, secondaryMode:secondaryMode, saveSlot:saveSlot});
+        trackEvent('game_start');
     } else {
       if (forcedScene) state.stats.sceneName = forcedScene;
       if (forcedStats) {
@@ -898,11 +906,21 @@ function redirectScene(sceneName, label, originLine) {
   clearScreen(function() {scene.execute();});
 }
 
+function restoreCheckpoint(slot) {
+  clearScreen(function() {
+    var scene = window.stats.scene;
+    delete scene.secondaryMode;
+    delete scene.saveSlot;
+    scene.restore_checkpoint(slot);
+  });
+}
+
 tempStatWrites = {};
 
 function transferTempStatWrites() {
   if (!_global.isIosApp) return;
   callIos("transferwrites", JSON.stringify(tempStatWrites));
+  tempStatWrites = {};
 }
 
 function getCookieByName(cookieName, ck) {
@@ -1131,6 +1149,10 @@ function updateSinglePaidSceneCache(sceneName, callback) {
       var url = canonicalHref + "scenes/" + fileName + "?hash="+hashes.scenes[fileName];
       xhr.open("GET", url);
       xhr.withCredentials = true;
+      if (window.beta) {
+        var betaPassword = window.betaPassword || "YmV0YTp5YXJkYXJt";
+        xhr.setRequestHeader("Authorization", "Basic " + betaPassword);
+      }
       xhr.onload = function() {
         var error;
         if (xhr.status !== 200) {
@@ -1195,7 +1217,7 @@ function updateSinglePaidSceneCache(sceneName, callback) {
 
 function updateAllPaidSceneCaches(receiptsSent) {
   if (!isStoreSceneCacheRequired()) {
-    if (window.isOmnibusApp && window.isAndroidApp) {
+    if (window.isAndroidApp) {
       window.receiptRequestCallback = submitReceipts;
       androidBilling.requestReceipts();
     }
@@ -1315,5 +1337,21 @@ function remoteConfig(variable, callback) {
   } else {
     var result = (_global.androidRemoteConfig && androidRemoteConfig.remoteConfig(variable)) || null;
     return safeTimeout(function() {callback(result);}, 0);
+  }
+}
+
+function trackEvent(name, params) {
+  try {
+    //console.log('trackEvent', name, params);
+    if (!params) params = {};
+    params.game_id = _global.storeName;
+    params.game_version = _global.version;
+    if (typeof nativeTrackEvent !== "undefined") {
+      window.nativeTrackEvent(name, params);
+    } else if (typeof gtag !== "undefined") {
+      gtag('event', name, params);
+    }
+  } catch (e) {
+    console.error("Error tracking event", e);
   }
 }